make Element a final class
